split mod
