+++ apps/forms/scaffolds/grid-editor/index.tsx
@@ -22,7 +22,6 @@ import {
   PieChartIcon,
   TrashIcon,
 } from "@radix-ui/react-icons";
-import { fmt_local_index } from "@/utils/fmt";
 import {
   Select,
   SelectContent,
@@ -31,12 +30,6 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Badge } from "@/components/ui/badge";
-import {
-  FormFieldDefinition,
-  FormResponse,
-  FormResponseField,
-  FormResponseSession,
-} from "@/types";
 import {
   DropdownMenu,
   DropdownMenuCheckboxItem,
@@ -50,85 +43,11 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { Button } from "@/components/ui/button";
-import clsx from "clsx";
-import type { GFRow } from "../grid/types";
 import { format, startOfDay, addSeconds } from "date-fns";
 import { format as formatTZ } from "date-fns-tz";
 import { LOCALTZ, tztostr } from "../editor/symbols";
-import type { DataGridFilterSettings } from "../editor/state";
-
-function rows_from_responses(responses?: FormResponse[]) {
-  return (
-    responses?.map((response, index) => {
-      const row: GFRow = {
-        __gf_id: response.id,
-        __gf_display_id: fmt_local_index(response.local_index),
-        __gf_created_at: response.created_at,
-        __gf_customer_uuid: response.customer_id,
-      }; // react-data-grid expects each row to have a unique 'id' property
-      response?.fields?.forEach((field: FormResponseField) => {
-        row[field.form_field_id] = {
-          type: field.type,
-          value: field.value,
-          storage_object_paths: field.storage_object_paths,
-        };
-      });
-      return row;
-    }) ?? []
-  );
-}
-
-function rows_from_sessions(
-  sessions: FormResponseSession[],
-  fields: FormFieldDefinition[]
-) {
-  return (
-    sessions?.map((session, index) => {
-      const row: GFRow = {
-        __gf_id: session.id,
-        __gf_display_id: session.id,
-        __gf_created_at: session.created_at,
-        __gf_customer_uuid: session.customer_id,
-      }; // react-data-grid expects each row to have a unique 'id' property
-      Object.entries(session.raw || {}).forEach(([key, value]) => {
-        const field = fields.find((f) => f.id === key);
-        row[key] = {
-          value: value,
-          type: field?.type,
-        };
-      });
-      return row;
-    }) ?? []
-  );
-}
-
-function applyFilter<
-  T extends FormResponse | FormResponseSession =
-    | FormResponse
-    | FormResponseSession,
->(
-  rows: Array<T>,
-  filter: DataGridFilterSettings,
-  datakey: keyof T = "raw"
-): Array<T> {
-  const { empty_data_hidden } = filter;
-  return rows.filter((row) => {
-    if (empty_data_hidden) {
-      if (row === null) return false;
-      if (row === undefined) return false;
-
-      const v = row[datakey];
-      return (
-        v !== null &&
-        v !== undefined &&
-        v !== "" &&
-        JSON.stringify(v) !== "{}" &&
-        JSON.stringify(v) !== "[]"
-      );
-    }
-    return true;
-  });
-}
+import { GridData } from "./data";
+import clsx from "clsx";
 
 export function GridEditor() {
   const [state, dispatch] = useEditorState();
@@ -160,16 +79,13 @@ export function GridEditor() {
 
   // Transforming the responses into the format expected by react-data-grid
   const rows = useMemo(() => {
-    switch (datagrid_table) {
-      case "response":
-        return responses
-          ? rows_from_responses(applyFilter(responses, datagrid_filter))
-          : [];
-      case "session":
-        return sessions
-          ? rows_from_sessions(applyFilter(sessions, datagrid_filter), fields)
-          : [];
-    }
+    return GridData.rows({
+      table: datagrid_table,
+      fields: fields,
+      filter: datagrid_filter,
+      responses: responses ?? [],
+      sessions: sessions ?? [],
+    });
   }, [datagrid_table, sessions, fields, responses, datagrid_filter]);
 
   const openNewFieldPanel = useCallback(() => {
